% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmprskHD.R
\name{cmprskHD}
\alias{cmprskHD}
\title{The main function to fit the Fine-Gray model with high-dimensional covariates}
\usage{
cmprskHD(surv, fstatus, id = 1:length(ftime), Z, failcode = 1,
  cencode = 0, inference.Z = 1:ncol(Z), init.tuning.method = c("cv",
  "1se", "1se.small"), init.glmnet.par = list(),
  inference.tuning.method = c("cv", "1se", "1se.small"),
  inference.glmnet.par = list(), var.est = c("initial", "onestep"),
  ncores = 1)
}
\arguments{
\item{surv}{vector of survival "\code{Surv}" objects with sojourn time and event indicator}

\item{fstatus}{vector with a unique code for each failure type and a separate code for
censored observations}

\item{id}{vector of subject ids (for time-depend covariates)}

\item{Z}{matrix of covariates}

\item{failcode}{code of fstatus that denotes the failure type of interest}

\item{cencode}{code of fstatus that denotes censored observations}

\item{inference.Z}{vector of the positions in the covariates for inference}

\item{init.tuning.method}{the tuning method for the initial LASSO estimator. Either
"cv", "1se" or "1se.small". "cv" is recommended}

\item{init.glmnet.par}{the list of parameters send to "glmnet" for the initial
estimator}

\item{inference.tuning.method}{the tuning method for the one-step-bias-correcting
estimator. Either "cv", "1se" or "1se.small". "cv" is recommended}

\item{inference.glmnet.par}{the list of parameters send to "glmnet" for the
one-step-bias-correcting estimator}

\item{var.est}{the method for the variance estimation. Either "initial", using the
initial estimator, or "onestep", using the one-step-bias-correcting estimator.
"initial" is asymptotically consistent while "onestep" is sometimes empirically better
for smaller sample size.}

\item{ncores}{number of cores for parallel computing}
}
\value{
a cmprskHD object
\item{initial}{The initial LASSO estimator}
\item{onestep}{The one-step-bias-correcting estimator}
\item{var}{The variance estimator for the one-step-bias-correcting estimator}
}
\description{
\code{cmprskHD}  returns the inital LASSO estimator, the one-step-bias-correcting estimator
and the variance estimator.
}
\details{
\code{fsurv} can have either 2 components for fixed covariates, sojourn time and status, or 3 components
for time-dependent covariates, entry time, exit time and status. The status in \code{fsurv} is coded
as "1" for the final entry of any type of observed event while "0" stands either for a non-final entry
or the final entry of an observed censoring. Similarly, the non-final entries should be coded as
the value of \code{cencode} in \code{fstatus}.
For the efficiency of the algorithm, we require that
the subjects are acsendingly sorted by their id, and
the entries for the same subject are consecutive and ordered in time from early to late.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
\author{
Marquis (Jue) Hou
}
