% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.phnet.R
\name{cv.phnet}
\alias{cv.phnet}
\title{Cross-validation for phnet}
\usage{
cv.phnet(x, y, ftime = NULL, fstatus = NULL, id, weights, offset = NULL,
  lambda = NULL, nfolds = 10, foldid, grouped = TRUE, parallel = FALSE,
  ...)
}
\arguments{
\item{x}{\code{x} matrix as in \code{phnet}.}

\item{y}{response \code{y} as in \code{phnet}.}

\item{ftime}{final time \code{ftime} as in \code{phnet}.}

\item{fstatus}{final status \code{fstatus} as in \code{phnet}.
Other causes are coded as integers starting from '2'.}

\item{id}{the subject \code{id} as in \code{phnet}.}

\item{weights}{observation weights. Default is 1 for each observation}

\item{lambda}{Optional user-supplied \code{lambda} sequence;
default is \code{NULL}, and \code{phnet} chooses its own sequence.}

\item{nfolds}{number of folds - default is 10.
Although nfolds can be as large as the sample size (leave-one-out CV),
it is not recommended for large datasets. Smallest value allowable is \code{nfolds=3}.}

\item{foldid}{an optional vector of values between 1 and \code{nfold} identifying what fold
each observation is in. If supplied, \code{nfold} can be missing.}

\item{grouped}{This is an experimental argument, with default \code{TRUE},
and can be ignored by most users. \code{grouped=TRUE} obtains
the CV partial likelihood for the Kth fold by subtraction;
by subtracting the log partial likelihood evaluated on the full dataset
from that evaluated on the on the (K-1)/K dataset. This makes more efficient
use of risk sets. With \code{grouped=FALSE} the log partial likelihood is
computed only on the Kth fold.}

\item{parallel}{If \code{TRUE}, use parallel \code{foreach} to fit each fold.
Must register parallel before hand, such as \code{doMC} or others. See the example below.}
}
\value{
An object with \code{S3} class \code{"cv.phnet"},
 which is a list with the ingredients of the cross-validation fit.
\item{lambda}{The values of \code{lambda} used in the fits.}
\item{cvm}{The mean cross-validated error - a vector of length \code{length(lambda)}.}
\item{cvsd}{The estimate of standard error of \code{cvm}. }
\item{cvup}{The upper curve = \code{cvm+cvsd}}
\item{cvlo}{The lower curve = \code{cvm-cvsd}}
\item{nzero}{The number of non-zero coefficients at each \code{lambda}. }
\item{phnet.fit}{The fitted \code{phnet} object for the full data.}
\item{lambda.min}{The value of \code{lambda} that gives minimum \code{cvm}.}
\item{lambda.1se}{The largest value of \code{lambda} such that error
is within 1 standard error of the minimum.}
\item{fordid}{The fold assignments used}
}
\description{
\code{cv.phnet}  Does k-fold cross-validation for phnet, produces a plot,
and returns a value for \code{lambda}.
}
\details{
The function runs \code{phnet} \code{nfolds+1} times;
the first to get the \code{lambda} sequence,
and then the remainder to compute the fit with each of the folds omitted.
The error is accumulated, and the average error and standard deviation over
the folds is computed.
Note that \code{cv.phnet} does NOT search for values for alpha.
A specific value should be supplied, else \code{alpha=1} is assumed by default.
If users would like to cross-validate \code{alpha} as well,
they should call \code{cv.phnet} with a pre-computed vector foldid,
and then use this same fold vector in separate calls to \code{cv.phnet}
 with different values of \code{alpha}.
 Note also that the results of \code{cv.phnet} are random,
 since the folds are selected at random. Users can reduce this
 randomness by running \code{cv.phnet} many times, and averaging the error curves.
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
\author{
Marquis (Jue) Hou
}
