% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phnet.R
\name{phnet}
\alias{phnet}
\title{The main function to fit the Fine-Gray model with high-dimensional covariates}
\usage{
phnet(x, y, ftime = NULL, fstatus = NULL, id = 1:nrow(x),
  family = c("cox", "finegray"), weights, offset = NULL, alpha = 1,
  nlambda = 100, lambda.min.ratio = ifelse(nobs < nvars, 0.01, 1e-04),
  lambda = NULL, standardize = TRUE, thresh = 1e-07, dfmax = nvars + 1,
  pmax = min(dfmax * 2 + 20, nvars), exclude, penalty.factor = rep(1,
  nvars), lower.limits = -Inf, upper.limits = Inf, maxit = 1e+05)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector.
Can be in sparse matrix format (inherit from class "\code{sparseMatrix}"
as in package \code{Matrix})}

\item{y}{a matrix with columns named 'time' and 'status', or 'start', 'stop' and
'status' for left-truncated data with time-dependent
covariates.
The last is a binary variable, with '1' indicating observed event (of the interested cause),
and '0' indicating right censored.
The function \code{Surv()} in package \code{survival} produces such a matrix.}

\item{ftime}{a vector of real values, the final time for each observations
under the Fine-Gray model with time-dependent covariates.}

\item{fstatus}{a vector of integer values, the type of cause for competing risks.
The cause of interest is coded as '1'.
Censoring is coded as '0'.
Other causes are coded as integers starting from '2'.}

\item{family}{Response type. Either "cox" or "finegray".}

\item{weights}{observation weights. Can be total counts if responses are proportion matrices.
Default is 1 for each observation}

\item{offset}{A vector of length nobs that is included in the linear predictor.
Default is NULL. If supplied, then values must also be supplied to the predict function.}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le \alpha \le 1}.
The penalty is defined as \deqn{(1-\alpha)/2\|\beta\|_2^2+\alpha\|\beta\|_1.}
\code{alpha=1} is the lasso penalty, and \code{alpha=0} the ridge penalty.}

\item{nlambda}{The number of \code{lambda} values - default is 100.}

\item{lambda.min.ratio}{Smallest value for \code{lambda}, as a fraction of \code{lambda.max},
the (data derived) entry value (i.e. the smallest value for which all coefficients are zero).
The default depends on the sample size \code{nobs} relative to the number of variables
\code{nvars}.
If \code{nobs > nvars}, the default is \code{0.0001}, close to zero. If \code{nobs < nvars},
the default is \code{0.01}.
 A very small value of lambda.min.ratio will lead to a saturated fit in the \code{nobs < nvars}
  case.}

\item{lambda}{A user supplied \code{lambda} sequence.
Typical usage is to have the program compute its own \code{lambda} sequence based
on \code{nlambda} and \code{lambda.min.ratio}. Supplying a value of \code{lambda} overrides this.
WARNING: use with care. Avoid supplying a single value for \code{lambda}
(for predictions after CV use \code{predict()} instead).
Supply instead a decreasing sequence of \code{lambda} values.
phnet relies on its warm starts for speed, and its often faster
to fit a whole path than compute a single fit.}

\item{standardize}{Logical flag for \code{x} variable standardization,
prior to fitting the model sequence.
The coefficients are always returned on the original scale.
Default is standardize=TRUE. If variables are in the same units already,
you might not wish to standardize.}

\item{thresh}{Convergence threshold for coordinate descent.
Each inner coordinate-descent loop continues until the maximum change in
the objective after any coefficient update is less than thresh times the null deviance.
Defaults value is \code{1E-7}.}

\item{dfmax}{Limit the maximum number of variables in the model.
Useful for very large \code{nvars}, if a partial path is desired.}

\item{pmax}{Limit the maximum number of variables ever to be \code{nonzero}.}

\item{exclude}{Indices of variables to be excluded from the model.
Default is none. Equivalent to an infinite penalty factor (next item).}

\item{penalty.factor}{Separate penalty factors can be applied to each coefficient.
This is a number that multiplies \code{lambda} to allow differential shrinkage.
Can be 0 for some variables, which implies no shrinkage,
and that variable is always included in the model.
Default is 1 for all variables (and implicitly infinity for variables listed in \code{exclude}).
Note: the penalty factors are internally rescaled to sum to nvars,
and the lambda sequence will reflect this change.}

\item{lower.limits}{Vector of lower limits for each coefficient;
default \code{-Inf}. Each of these must be non-positive.
Can be presented as a single value (which will then be replicated),
else a vector of length \code{nvars}.}

\item{upper.limits}{Vector of upper limits for each coefficient; default \code{Inf}.
See \code{lower.limits}.}

\item{maxit}{Maximum number of passes over the data for all \code{lambda} values;
default is \code{10^5}.}
}
\value{
An object with \code{S3} class \code{"phnet"}.
\item{call}{the call that produced this object}
\item{beta}{a \code{nvars} by \code{length(lambda)} matrix of coefficients,
stored in sparse column format (\code{"CsparseMatrix"}). }
\item{lambda}{The actual sequence of \code{lambda} values used. When \code{alpha=0},
the largest \code{lambda} reported does not quite give the zero coefficients reported
(\code{lambda=inf} would in principle). Instead, the largest \code{lambda} for \code{alpha=0.001}
 is used, and the sequence of \code{lambda} values is derived from this.}
\item{dev.ratio}{The fraction of (null) deviance explained.
The deviance calculations incorporate weights if present in the model.
The deviance is defined to be 2*(loglike_sat - loglike),
where loglike_sat is the log-likelihood for the saturated model
(a model with a free parameter per observation). Hence dev.ratio=1-dev/nulldev.}
\item{nulldev}{Null deviance (per observation). This is defined to be
2*(loglike_sat -loglike(Null)); The NULL model refers to the 0 model. }
\item{df}{The number of nonzero coefficients for each value of \code{lambda}.}
\item{dim}{dimension of coefficient matrix (ices).}
\item{nobs}{number of observations.}
\item{npasses}{total passes over the data summed over all \code{lambda} values.}
\item{offset}{a logical variable indicating whether an offset was included in the model.}
\item{jerr}{error flag, for warnings and errors (largely for internal debugging).}
}
\description{
\code{phnet}  fit a Proportional Hazards model with lasso or elasticnet.
}
\details{
\code{y} can have either 2 components for fixed covariates, sojourn time and status, or 3 components
for time-dependent covariates, entry time, exit time and status. The status in \code{y} is coded
as '1' for the final entry of observed events from the cause of interest and as '0' otherwise.

With the time-dependent covariates, all entries exit prior to the final sojourn time
are marked as '0' for \code{fstatus}. Entries of a subject with a competing risks event
enter after the final sojourn time should be included in the data and are marked as
codes greater than '2'.
The \code{ftime} of these entries should be the associated competing risks event times.
The 'stop' in \code{y} for these entries should not mark the competing risks event times
but rather the covariate-change times.
Please code 'stop' in \code{y} for the last entry of a subject with the competing risks event
as the maximal event time.
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
\author{
Marquis (Jue) Hou
}
