% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.phnet.R
\name{predict.phnet}
\alias{predict.phnet}
\alias{coef.phnet}
\title{Make predictions from a "phnet" object.}
\usage{
\method{predict}{phnet}(object, newx, s = NULL, type = c("link", "response",
  "coefficients", "nonzero"), exact = FALSE, offset, ...)

\method{coef}{phnet}(object, s = NULL, exact = FALSE, ...)
}
\arguments{
\item{object}{Fitted \code{phnet} matrix as in \code{phnet}.}

\item{newx}{Matrix of new values for x at which predictions are to be made.
Must be a matrix; can be sparse as in \code{Matrix} package.
This argument is not used for \code{type=c("coefficients","nonzero")}.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions are required.
Default is the entire sequence used to create the model.}

\item{type}{Type of prediction required. Type \code{"link"} gives the linear predictors.
Type \code{"response"} gives the fitted relative-risk.
Type \code{"coefficients"} computes the coefficients at the requested values for \code{s}.
Type \code{"nonzero"} returns a list of the indices of the nonzero coefficients
for each value of \code{s}.}

\item{exact}{This argument is relevant only when predictions are made at values of
\code{s} (lambda) different from those used in the fitting of the original model.
If \code{exact=FALSE} (default), then the predict function uses linear interpolation
to make predictions for values of \code{s} (lambda) that do not coincide
with those used in the fitting algorithm.
While this is often a good approximation, it can sometimes be a bit coarse.
With \code{exact=TRUE}, these different values of \code{s} are merged (and sorted) with
\code{object$lambda}, and the model is refit before predictions are made.
In this case, it is required to supply the original data \code{x=} and \code{y=} as
additional named arguments to \code{predict()} or \code{coef()}.
The workhorse \code{predict.glmnet()} needs to update the model,
and so needs the data used to create it. The same is true of
\code{weights, offset, penalty.factor, lower.limits, upper.limits}
 if these were used in the original call. Failure to do so will result in an error.}

\item{...}{This is the mechanism for passing arguments
like \code{x=} when \code{exact=TRUE}; see \code{exact} argument.}

\item{newoffset}{If an offset is used in the fit, then one must be supplied
for making predictions (except for \code{type="coefficients"} or \code{type="nonzero"}).}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods,
this functions predicts fitted values, logits, coefficients
and more from a fitted \code{"phnet"} object.
}
\seealso{
\code{\link[glmnet]{predict.glmnet}}
}
\author{
Marquis (Jue) Hou
}
